Update: Se creo la tabla parametros para almecenar parametros(IVA), se hicieron sus casos de uso(solo modificar), se hizo la migración y en pedido se cambio el tipo de parametro que recibe PrecioFinal.

+++ LogicaNegocio/EntidadesNegocio/Pedido.cs
@@ -77,7 +77,7 @@ namespace LogicaNegocio.EntidadesNegocio
             return total;
         }
 
-        public decimal PrecioFinal(double iva)
+        public decimal PrecioFinal(decimal iva)
         {
             decimal precioTotal = PrecioTotalLineas();
             if(TipoPedido == TipoPedido.PEDIDOEXPRESS)
@@ -90,7 +90,7 @@ namespace LogicaNegocio.EntidadesNegocio
             if (TipoPedido == TipoPedido.PEDIDOCOMUN && Cliente.DistanciaPapeleria > 100)
                 precioTotal *= (decimal)0.05;
 
-            return precioTotal * ((decimal)iva/100);
+            return precioTotal * (iva/100);
         }
 
     }
